+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+[0.6.0] - 2022-08-02
+
+### Added 
+
+- Unit tests for awsbatch.py.
+
+### Removed
+
+- Test action for Python Version 3.10.
+
+## [0.5.0] - 2022-08-03
 
 ### Changed